Added CMake logic to enable linking against libzerotiercore.a
Configure with -DNNG_ENABLE_ZEROTIER=yes -DNNG_ZEROTIER_SOUCE=<path>
The <path> must point to a dev branch, checked out, and built with
a libzerotiercore.a in the top directory, and a ZeroTierOne.h header
located at include.  The build will add -lc++ to the compile, as
the ZeroTier core functionality is written in C++ and needs some
runtime support (e.g. new, delete, etc.)