We are now properly initializing (it seems!) zerotier.
We can see identity files created, the planet file exists, and
messages are exchanged with the outside world (VL1 stuff).

We notice that the files are created with incorrect (too open)
permissions, and we have some issues with creating directories
(the application program must ensure that the home dir already
exists).   We should begin the work for a portable file system
layer.

It looks like its also time to start exchanging our ethertype 0x901
messages.