Snapshot progress.  We have just passed our first virtual frame!
So it takes a while for the virtual network configuration to settle
down, and if you don't wait long enough the core will just drop the
frame.  We will need to start the work to set up for retries so that
end users do not need to worry about this.  Also, given that ZeroTier
can take a while for the network to establish, we will want to set up
an AIO or something and adjust the timers for connection.  Clearly 5s
in inadequate if we're still configuring the network.

While here we fixed the state routines which were clobbering the files
on every open ("wb" instead of "rb"), so now we should see greatly
reduced node creation counts.