Create ephemeral nodes by default.
This creates an ephemeral node, which means apps don't need to set
the ZeroTier home path in order to use ZeroTier.  Persistence is only
required if needed by the app.  This may facilitate use in settings
where persistence is untenable (completely stateless devices for
instance).

All endpoints will by default share the same ephemeral node.

This likely means that creating an endpoint the first time may take
slightly longer, as we have to generate key material on startup.
This only applies to the first time an endpoint is created, however.