Use the listen port for the entire new connection.
This eliminates the step where the server supplies a new randomly
chosen port, instead using the same port that the service was
originally contacted on.  This should make it much easier to write
stateless filtering rules, and may help folks doing debug to
understand the ports being contacted. Packet dumpers can also use
this to perhaps perform more meaningful decodes in the future.